Reuse Anub'ar Ambusher's effect for Ancient Brewmaster.
