+++ src/main/java/com/hearthsim/card/minion/concrete/AncientBrewmaster.java
@@ -5,11 +5,14 @@ import java.util.EnumSet;
 import com.hearthsim.card.minion.BattlecryTargetType;
 import com.hearthsim.card.minion.Minion;
 import com.hearthsim.card.minion.MinionTargetableBattlecry;
+import com.hearthsim.event.EffectMinionBounce;
 import com.hearthsim.exception.HSException;
 import com.hearthsim.model.PlayerSide;
 import com.hearthsim.util.tree.HearthTreeNode;
 
 public class AncientBrewmaster extends Minion implements MinionTargetableBattlecry {
+
+    protected final EffectMinionBounce effect = new EffectMinionBounce();
 
     public AncientBrewmaster() {
         super();
@@ -25,16 +28,9 @@ public class AncientBrewmaster extends Minion implements MinionTargetableBattlec
      */
     @Override
     public HearthTreeNode useTargetableBattlecry_core(PlayerSide side, Minion targetMinion, HearthTreeNode boardState) throws HSException {
-        HearthTreeNode toRet = boardState;
-        if (toRet != null) {
-            if (boardState.data_.getCurrentPlayer().getHand().size() < 10) {
-                Minion copy = targetMinion.createResetCopy();
-                if (copy != null) {
-                    toRet.data_.getCurrentPlayer().placeCardHand(copy);
-                }
-            }
-            toRet.data_.removeMinion(targetMinion);
+        if (boardState != null) {
+            effect.applyEffect(PlayerSide.CURRENT_PLAYER, this, side, targetMinion, boardState.data_);
         }
-        return toRet;
+        return boardState;
     }
 }
